// src/core/types.ts
/**
 * 颜色枚举：红方和黑方
 */
export enum Color {
  Red = 'RED',
  Black = 'BLACK',
}

/**
 * 棋子类型枚举
 */
export enum PieceType {
  General = 'GENERAL',   // 将/帅
  Advisor = 'ADVISOR',   // 士/仕
  Elephant = 'ELEPHANT', // 象/相
  Horse = 'HORSE',       // 马
  Chariot = 'CHARIOT',   // 车
  Cannon = 'CANNON',     // 炮
  Soldier = 'SOLDIER',   // 兵/卒
}